More playing around
I've introduced a mode in which we can distinguish what we draw. Ideally
this will become a separate widget (maybe?) which shows the list of
mails and the current selected mail with header and bodies.

Interestingly creating mails from the notmuch query resulted in a
segfault, so I've currently just used placeholders there.