Allow to edit the search terms
This adds a little input box at the bottom of the main window to allow a
user to edit the notmuch search. This is still all very much proof of
concept.

I did lump in some additional changes which is the fact that widgets
should be named. For a growing application there is no way we will be
able to get around it.

The mails found are currently only updated if the search terms
change (and the user presses enter).

I've played around with the change of focus for the different widgets on
the main screen. That is, if the user presses ':' - I've leaned towards
the vim behavior - the input box receives all keyboard events. This is
currently done by tracking modes in the application state. Not sure if
that is sensible for the future.