Refactor event handlers into (some) Keybindings
In order to make the application keyboard handlers customizable,
refactor most handlers into a form of keybinding. This excludes default
handlers for the widgets we use and just forward to the widget themself.
There is possibly a need for also customize those in the future, but
will have to be addressed by a separate patch.

Issue: #27