Compose e-mail
This implements a very basic way of sending e-mails. The editor being
used to edit the e-mail body is whatever is set in $EDITOR and falls
back to vim. Perhaps that will be changed down the line by simply
configuring your favorite editor.

At this point in time only plain text e-mails can be sent,
since the editor to edit attachments is not fully functional yet (see #10).

This interestingly also implements a way in regards to Issue #8. The
current keybinding to toggle between the index view and composing an
e-mail is TAB. So the user can switch back to the index while writing
his e-mail.

Caveats:

The cycling through the input widgets is currently not very nice.
Perhaps all this can be made easier by using Brick.Focus

Issue: #13