Show formatted date in Index
This shows the formatted message date in the mail index. The format is
currently hard coded. Suppose we can add the configuration option later.

One particular reason to not spend the time now is that I think it will
probably make sense to create a parser so to make the entire list item
configurable. How this might look for now, I don't know.