Xmonad style config
This now actually implements the XMonad style configuration. It is
achieved with some re-exports of Purebreds dependencies including a
small set of lens operators.

The configuration is typically looked up under the user config directory
in $HOME/.config and the system GHC is used to compile the
configuration. This can be overwritten however with two variables
respectively:

  WITHSTACK=1 to indicate using 'stacks' ghc
  PUREBRED_CONFIG_DIR=<filepath> to indicate using a non-standard
  configuration directory

Issue: 8