Use an 'IO FilePath' as config to get db dir
We want to provide a sensible default DB path by deriving it from
the notmuch config file.  This requires IO.

We also want users to be able to override the DB path (typically
using a pure value).  To keep things simple for users, make the
default configuration contain an 'IO FilePath' - an IO computation
for getting the db dir.  Then the first step when running the
program is to process that into an "internal" configuration value.
Hence we abstraction the Configuration type over the type of the db
file path config.