Do not access an un-indexed field
Accessing a notmuch un-indexed field results in notmuch parsing the
mail. This results in an open file descriptor which in turn ends up
open more FDs than the GC can close. The end result is during setting up
the terminal, vty needs to read the terminfo file which the system
denies.

Issue: #49