Build hs-notmuch against newer notmuch
This installs notmuch locally since the notmuch version available via
system packages is very, very outdated. Furthermore it has a regression
in a macro on which hs-notmuch is dependend upon. Result is failing tests.

To circumvent this problem, install our own supported version of notmuch
and build hs-notmuch against this version.