tests: Fixes intermittently breaking test
This fixes an intermittently breaking test, which has a timing problem.
The test driver waits for a string which is already present on the
screen capture which results not in waiting for a re-draw. Since the
body is not shown when all headers are visible, we use the string in the
mail body to wait for to circumvent the timing problem.