Refactor Keybindings
Split up the Action in to it's own datatype from the binding itself.
The motivation is to compose multiple actions if there is need.

For example, to move the cursor up and down the index we can define one
action: moveIndexUp. Yet we can re-use the same action with an action to
update the current view with a parsed mail when we actually view
e-mails.

Issue: #45