Change event handlers to use Vty.Event directly
Event handlers are currently dealing with BrickEvent, but in fact we
are only ever actually using the VtyEvent case of that type.  This
means that some event handlers are doing unnecessary case analysis.

Push the case analysis up into 'appEvent' and from that point on
just pass Vty Event values to the sub-handlers.