Fix tests
In the CI environment the tests can be a bit slow, so we have to be
certain that the condition we are waiting for in the output is not
already satisfied.  In this case, when opening a mail we were
looking for the subject, but the subject was already present in the
BrowseMail view.

Instead look for part of the body of the email.  The name of the
view in the status bar also provides a nice gate to ensure that the
"next view" has been rendered before executing further tests against
the terminal content.