Fix compose editor visually
Initialise the editors for composing a new mail only using one line
instead of multiple lines. Restrict the rendering to one line too.

While it might make sense to hold multiple lines of text, we currently
want to restrain them to a single line in order to render them neatly in
the interface.