ci: remove unneeded ghc packages
We are using stack which can install its own copy of GHC.  The
version of GHC will drift over time for stackage nightly, which
makes this a bit of a maintenance burden.  The GHC installed by
stack gets cached, so just let stack worry about GHC instead of
specifying it manually.