tests: use "keepalive" tmux session and unique session names
We have experienced problems with the `tmux new-session` command:
a race of server availability when the server is being torn down
after the previous session gets killed.  Create a "keepalive"
session whose lifetime is the whole of the acceptance tests, to
prevent the server from being killed.

Also sequentially number the test sessions to avoid any potential
races when sessions are being killed and started.

Fixes: https://github.com/purebred-mua/purebred/issues/123