switch tag parsing back to attoparsec
We want to avoid depending on two different parser libraries.
So switch back to attoparsec for parsing tags.

This refactor also changes Purebred to use the Notmuch.Tag type
internally.  User input is parsed into a Notmuch.Tag during parsing.
Anything that will not be accepted by libnotmuch is rejected at that
point, so there is now one less error that can occur later on when
trying to write tags.

attoparsec's error handling is not as nice as parsec so add a couple
of helper parsers for getting nicer errors out of it.  The error
message did change a little, so update the test accordingly.

An orphan Arbitrary instance is added for Notmuch.Tag.  The Tag is
of a random length (between 1 and the max length), and consists of
random printable ASCII characters excluding SPACE.  Bump the
hs-notmuch dependency to a commit where 'tagMaxLen' is exported (we
need it for the Arbitrary instance).