Configuration: add type synonym for lenses to non-abstract fields
Each time we add a new abstract field in configuration, we have to
add a type parameter to 'Configuration'.  Then we have to chase that
change at all the optics.

But for non-abstract fields, we can avoid that work by defining a
single type synonym.  Then only the type synonym needs to be
updated.