Test the configuration
This is not only useful for shipping working example configurations, but
also needed for further acceptances tests. For testing sending mail for
example, I foresee changing the send command (new config option) to
writing the mail to the local file system instead of giving it to
/usr/bin/sendmail. Best idea I could come up with, with a test suite
which can still be run as an integrated part rather than separate.

Fixes https://github.com/purebred-mua/purebred/issues/41