Allow to customize the sendmail function
Instead of hard coding how we send the mail via send mail, set a default
function in the configuration.

This is also a first step to support writing a copy of the mail to the
local file system, before/after sending it (see #119).

Furthermore, this can be used in the user acceptance tests to test
sending an e-mail. Arguably this doesn't actually test the sending, but
comes as close as possible with the current setup.

For expanding the testing I think we'd need a way to catch the mail
being sent through sendmail (see #156).