Upgrade to lts-10.5
Note, the jump to brick 0.34 is warranted because 0.32 introduced forms which we
want to use for the compose editor, 0.33 had additional validators for form
fields which we need as well. The 0.34 version has API changes which we use
throughout the code and better adopt/fix in cases of misalignment earlier than
later.