implement charset decoding
Update to commit of purebred-email that introduces charset decoding
and modify purebred to perform charset decoding when displaying a
message.

Drive-by generalisation of getHeader, getFrom, getSubject and getTo
which work for all instantiations of 'Message s a'.

Also do a drive-by addition of a colon between header name and
header value in the mail display widget.