Keep tags to the left
The placement of the tags between threads and mails was inconsistent. They were
left in the threads, but placed to the far right in the list of mails. The
benefit of having them placed left is that the remaining space can be used for
the subject, which tends to be naturally longer than the tags.

This patch now creates a consistent placement between the views.