fix up composition
Rendering can cause a run time error, since the renderer caught the composeFrom
editor twice. That happened by using the same function to render the focused
view widget when going over the list of names to render. The second instant was
that the replaced ComposeFrom editor remained in the Threads view and was never
replaced by the default again.