remove class EventHandler
Remove a layer of unnecessary indirection: class EventHandler.
Define an EventHandler data type instead.  The calling code is much
more compact now.

Also add the 'nullEventHandler' EventHandler which does nothing.  In
the future it could be enhanced (or a similar handler added) to
trigger a message that the key was not bound.