fix mail view event dispatch
If viewing a mail from thread list, backing out to mail list, then
selecting a mail, we end up in ViewMail/ListOfMails rather than
ViewMail/ScrollingMailView.  The former is not mapped to an event
handler.

I'm not entirely sure why we end up in that mode, so for now just
dispatch to ScollingMailView keybindings whenever ViewMail is the
active view.