tidy up composite action descriptions
The "noop" action and the currrent means of describing composite
actions - writing "... and then ..." - led to some pretty ugly
descriptions.  Tidy it up by preserving descriptions as a (possibly
empty) list of strings, and combining them in the obvious way.  Make
the output a bit more compact: the combining text is now " > "
instead of "and then".

Also change the type from String to Text for efficiency.