Remove to compose mails from mail list
Removed keybinding from the list of mails, since it will just get the
user stuck. Typically users will always land on the view which shows the
mail first. So to compose a mail from the list of mails, you'd have to
press two keys to intentionally start composing. I don't think it's
worth it.

In order to support this, we'd need to add quite a bit of new code for
additional keybindings and drawing. Since we're currently not really
sure if we want to really keep the way we're composing mails, rather
just support the mail composition from the thread view.