Show if editor input can not be parsed
When entering input (tagging, addresses) Purebred now gives instant feedback if
the input can be parsed or not. The error message itself is not very
helpful, since it is the raw parsing error. However it's a start for a
more versatile instant feedback system.

The implementation forks a thread which sends a PurebredEvent every X
seconds. As long as the user keeps on entering characters, Purebred will
kill the existing thread and spawn a new one. Once the user stops
entering characters, Once the focus is removed, the thread for
validating the current input is not killed and sends the error (if any)
as an application event. The event with the error (if any) is then set
in the application state and displayed.

Fixes https://github.com/purebred-mua/purebred/issues/298