Poc: use cache for rendering list items
This uses an idea discussed on #208 by Christopher Lewis to improve
scrolling performance on lists. This idea assigns a name to each list
item in order to cache it.

This results in a much faster rendering. However, this patch does not
add cache invalidation which results in no update of the list item if
the user applies action to one. For example, adding/removing tags to
threads will not show the change list of tags.

TODO:
* use a better architecture to invalidate cached list items