Adds reply to mail feature
This "fixes" the reply to mail action, which was never correctly implemented due
to missing support in the email parsing library. Using purebred-email, there is
wide support for manipulating the mail body in order to properly quote the body.

This is the first take on the replying action and will only work for text/plain
content types.

Fixes https://github.com/purebred-mua/purebred/issues/114