lazily load threads
If search result is very long, it takes a long time to load all the
items.  Preliminary work in brick, hs-notmuch and purebred leads to
what is now a small change to enable lazy loading of search results.

First, we move to (Haskell) notmuch-0.2 which fixes issues with lazy
result loads.  In particular, derived objects always hold a
reference to their parent object, to prevent it being GC'd, which
caused segfaults when the child object (wihch is an opaque pointer)
held pointers into the parent.  One programmer-friendly consequence
of this is that resource management of the database handle is now
implicit.  The database will be closed when it gets GC'd, which will
only happen when all derived objects (queries, threads and messages)
have been GC'd.

Second, we update the thread index list to use the 'V' container
type added a little while ago, which supports lazy loading.

Third, for processing notmuch Thread objects into purebred's
NotmuchThread type, we define a []- and IO-specific 'lazyTraverse'
function that uses 'unsafeInterleaveIO' to retain the lazy loading
behaviour afforded by hs-notmuch.

Finally, we tweak the functions that update the AppState to not set
the 'length' of the list (because 'length' would force the entire
spine of the container).

Fixes: https://github.com/purebred-mua/purebred/issues/233