tests: Exponentially back off waiting
This patch adds an exponential back off to the wait times when waiting
for a repaint of the UI. With the default of 5 attempts we increase the
wait time exponentially by 4:

>>> 20000 / 1e6
0.02
>>> _ * 4
0.08
>>> _ * 4
0.32
>>> _ * 4
1.28
>>> _ * 4
5.12

cummulating to 1.2s for the UI to be repainted for every step. This is
shorter to the previous setting which waited for every step 0.5s which
was cumulatively 2.5s until it bailed out.

This has the potential to shorten the test run time from 98s down to
30s.

Note: The UI repaint depends on the IO rate of the machine the tests
run. No consideration has taken into account to determine the IO rate
for an optimal initial back off time with this patch.