reset list when applying search
When applying a search we currently try and preserve the currently
selected item.  This doesn't make a lot of sense - it will usually
be a different item.  Furthermore when using a lazy list type it
will force evaluation of the result up to the current index
(undesirable).

Therefore, when applying a search always set the selected index back
to zero.