Revert "tests: Exponentially back off waiting"
This reverts commit cb94323f8d7442e5f59f4169afc38b388d2ca61e.

The more aggressive (read shorter) initial timeouts could lead to more
failing tests. This is partially because after we've implemented the new
UI methods, we never investigated that the views now share more common
widgets. Widget labels show up in more different views. This leads to
more failures in combination with the aggressive timeouts, because the
assertions still find the same strings in the old views as well as in
the new views.

Therefore, tweaking more with the timeouts will make the tests more
brittle.