uat: tell capture-pane to unwrap wrapped lines
When sending strings into a tmux session, if the lines wraps in the
terminal, capture-pane will return the _wrapped_ output, newlines
and all.

With long environment variables and (when running tests locally) a
long command prompt, the line can wrap and confuse the condition
checker.  For example:

  sending mail successfully:                           FAIL (5.04s)
    Wait time exceeded. Condition not met:
        'Literal "export PUREBRED_CONFIG_DIR=/tmp/purebredtest10834"'
         last screen shot:

    [T470s:~/dev/purebred] [ ci/avoid-stack● ] ftweedal% export TERM=ansi
    [T470s:~/dev/purebred] [ ci/avoid-stack● ] ftweedal% export PUREBRED_CONFIG_DIR=
    /tmp/purebredtest10834

    <lots of blank lines (expected)>

     raw: <snip>export PUREBRED_CONFIG_DIR=\n/tmp/purebredtest10834<snip>

Observe the unexpected "\n" in the middle of the line.  (The fact
that this occurs right after an '=' is mere coincidence.

tmux capture-pane provides the '-J' option for unwrapping lines
(even though they are being displayed wrapped).  Use this option to
avoid confusing the condition checker.

This change requires tweaking a couple of the regex conditions due
to the presense of additional whitespace in the captured pane.