uat: share config dir and binary across test cases
All our test cases do (or can) use the same configuration.  Right
now, we are incurring the recompile cost for every test case.  But
we only need to incur it once, and just point the program to a fresh
maildir for each test case.  This saves some time in the test suite.

This commit also resolves another problem: if the config dir was not
explicitly specified (via the PUREBRED_CONFIG_DIR environment
variable) and the tests are being run by a user who already had a
purebred configuration, the _user's_ purebred configuration would be
used.  The mail directory was overridden, but the user's custom
binary could have, for example, incorrect keybindings that will
cause tests to fail.  Explicitly setting the config dir for every
test session averts this problem.

In terms of implementation, config dir setup and cleanup moves to
'main', and a new 'GlobalEnv' data type conveys the details to the
individual test cases.

The ability to use a custom configuration for a single test case is
provided for, but not used.  The only change that would be required
is to enhance 'withTmuxSession' to provide the datum to 'setUp', and
'setUp' to actually create it and set 'envDir' field in the session
'Env'.  The cleanup side of things is already implemented.