tests: include test data in source distribution
The test files are currently not included in the source
distribution.  Include them.

To avoid having to write out a bunch of very long filenames for the
contents of Maildir/, make use of wildcard syntax and behaviour
available since cabal 2.4.  Update the package description
accordingly.  This entailed moving the "cabal-version" declaration
to the top of the file, updating the "license" field to include an
SPDX identifier (see https://spdx.org/licenses/), and adding the
'autogen-modules' declaration.

Also drive-by update the "copyright" declaration to include the year
2018.