uat: fix "application has quit" condition
When the application quits and the shell writes out the command
prompt, the cursor might not be in column 0.  This could cause the
target string to be broken across lines, which leads to strange
results when ANSI escapes for colour/style are involved.

We can avoid these issues by pressing Enter again, to ensure that
the shell prompt gets written again from column 0, and by setting
PS1 to a relatively short string.  But we have another problem -
purebred has been sent the 'q' key to quit, but if the keys are sent
all at once, there's a very good chance purebred is still running
and eats the keypress events (which disappear into the aether).

To work around these issues:

- Set PS1 to a string that is otherwise unlikely to appear in the
  purebred output, and

- After sending 'q', wait a little bit for purebred to terminate,
  then send 'Enter' to write out the prompt again before awaiting
  the expected output