uat: make filebrowser tests handle different files
If for whatever reason the directory in which purebred was run does
not contain certain files which are currently hardcoded, the tests
fail.  This can happen if you are running tests from a source
distribution instead of a git checkout (i.e. the sdist may omit some
files).  Make the test more resilient by dynamically determining the
contents of the directory in which purebred is running.