use correct notmuch SearchTerm constructors
The 'FreeForm' constructor is not correct.  Although the current
'show' instance does not quote it, this is a bug.  The correct
constructor for passing a query string to libnotmuch as-is is Bare.
Change to the correct contstructor and also access them via the
qualified Notmuch import (Notmuch.Search will be removed in a future
release).