lay groundwork for custom event processing
brick supports custom events.  In time we will use custom events,
supplied to the event loop via a 'BChan', for various purposes.
This commit lays some groundwork for the use of custom events.  In
particular:

- the PurebredEvent type is defined (with no constructors, for now)

- the app event handler has an equation for processing PurebredEvent
  events (an empty case, for now)

- during initialisation, a BChan is constructed and stored in the
  InternalConfiguration

Additionally, drive-by refactor the Configuration data type so that
the "boundary" field is only present in the InternalConfiguration
(it makes no sense to expose it to users).  Also, simplify several
lens definitions for the Configuration data type.