fork thread to compute length of thread index
After performing a notmuch search, fork a thread to compute the
length of the result (fully evaluating the spine of the list) and
send a notification to brick to update the model.

This adds the 'NotifyNumThreads' constructor to the 'PurebredEvent'
type, and updates the event handler to process it.

The 'initialState' function is refactored to reuse 'applyThreads',
where the "fork background thread" behaviour is defined, so that the
thread is spawned for the initial search upon startup, too.

Part of: https://github.com/purebred-mua/purebred/issues/233