ignore obsolete NotifyNumThreads events
If a search with a small result is performed after a search with a
large result, the NotifyNumThreads event for the earlier search can
arrive after the NotifyNumThreads event for the later search.  As a
consequence, the wrong number of items will be displayed.

Define the 'Generation' data type for recording a particular
"generation" of the 'AppState' (or some aspect thereof).  Events can
also carry a 'Generation' to indicate the generation(s) of the
AppState to which it is related.

Update the 'NotifyNumThreads' event data type and the application
state for the thread index to each include a generation field.
Update the event handler to ignore 'NotifyNumThreads' events that do
not match the current generation of the thread index.

Part of: https://github.com/purebred-mua/purebred/issues/233