move event types to new module
Add the Purebred.Events module and move types and functions related
to events to it.

Update the generation 'Generation' to hide the constructor and avoid
the 'Enum' instance.  This avoids a conflict between
-XDeriveAnyClass and -XGeneralizedNewtypeDeriving, which caused a
build failure with GHC 8.0 and warnings with later versions.  The
removal of the 'Enum' type class also lets us enforce monotonicity.