Extend the data type for drawing/hiding
We used two data structures to determine the view state of the widget and the
focus state: a simple list and a focus ring. The problem with the setup was when
we needed to adjust the focus and the view state. The focus ring does not
provide any operations to perform on the items at all, neither does the circular
list type. Keeping what to draw in a separate list also means trouble, since
focus ring and list of widgets could get out of sync.

Instead use a simple setter to determine which widget has the focus. The patch
introduces an additional data type to determine which widget is visible or
should be hidden from the UI. Because of the short comings of FocusRing and
circular list, we use lenses for accessing the data types in order to change
their view states.

That allows us to operate on widgets and their view state more reliably.

The only draw back is that we need to "find" the next focusable and or visible
element. Given that our list of widgets to render is small I think is
acceptable.

Fixes https://github.com/purebred-mua/purebred/issues/181