CI: test our Nix setup
This patch adds additional items to our build matrix testing building with nix.

We keep the default before and install scripts for the cabal tests, but
overwrite them with nil values specifically for the nix language items.

Kudos to
https://github.com/gibiansky/IHaskell/commit/d5c6272390f8eaf5260bf593414ff70b1c362d35
which I had found with a random search and used as a reference.