Move view related functions into the view module
This is basically just to make view related functions consistent with the views.
One slight change made on how we determine the focused widget is due to changes
from the focus ring. Since we now have a definitively focused widget, we don't
need to pass in a fallback widget any more.