sanitise control characters in message bodies
Control characters in message bodies can result in UI corruption
(e.g. when horizontal tab characters occur).  It is also a security
issue (ANSI escapes causing undesired terminal behaviour).

Add a Text sanitisation function and use it when converting entities
for display.

There are most certainly other places in the UI we need to employ
it, and we should properly audit the codebase - perhaps even using
phantom types to taint unsanitised external data.  But this is a
starting point.

Also do a drive-by tidy-up of UI.Utils.

Fixes: https://github.com/purebred-mua/purebred/issues/265