Delete deprecated code
This code was kept to achieve the same with a Vector data structure (since it
preserves order) than what we had with a focus ring before. Due to the fact that
the UI resembles more mutt keybindings and behaviour, composing mails was not
behaving consistently to the rest of the UI. Most specifically that the user had
to actively move the focus to the next editors in order to adjust the
headers (see #263).

With the change of removing this inconsistency, there is no need to keep this
awkward code any more.