Allow to automatically open attachments
If the lookup for a content type is successful, we use the mapped command,
otherwise show an input editor. Keep the possibility to open the attachment with
a different command by binding the previous open action to a different
keystroke.

A sample mailcap is used and adds a mapping from text/html to elinks and a catch
all to xdg-open. This patch would also fix #174 since it is now possible to
automatically invoke a program to view an HTML attachment.

Fixes https://github.com/purebred-mua/purebred/issues/182 and
      https://github.com/purebred-mua/purebred/issues/174

erasf