Make Mailcap entry aware of temporary files
When opening attachments with mailcap entries, it is sometimes possible that the
actual binary just sends a message to the main program and exits. Deleting the
temporary file upon process exit would cause the main program upon opening the
temporary file to fail since the file would have been deleted by that time.
Sometimes opening the file from the main program just works, because the time
difference between exiting and opening is really quick, but mostly it doesn't.

Instead, make the mailcap entry flexible so that depending on the command we
either keep or delete the temporary file.

Fixes https://github.com/purebred-mua/purebred/issues/276