Drop the use of Themes
The current implementation did not fully utilise the Theme support of Brick. In
fact we've converted the theme with each redraw into an AttrMap, which probably
was costing performance points.

In order to make use of it, we'd need to support loading the theme config files
so we allow users to adjust the themes. Two problems I found which led to
dropping this for now:
a) the INI parser is too finicky: it's white space sensitive, which means
missing a new line after your last field value renders the INI useless
b) the customisations only overwrite existing attributes. That makes it harder
to support setting attributes which are user specific and not set by default.