Refactor IO error handling for attachments
With the refactoring for handling temporary files (see
dd8ad0d062a418c3bc09ef63c6de36b5b5af908a) we introduced a slightly
different way to handle errors. This is slightly different to the way
we've handled errors from IO computations before.

Throughout most of our code, we use MonadError in order to transform and
propagate errors in IO computations.

This patch changes the error reporting we used to be consistent to what
we have throughout the code base. Most types have changed to use
MonadError as well as functions in order to propagate the error to the
user interface.

This becomes especially important when we further make changes to the IO
code which needs extensive error handling and propagation to the user
interface.