New action to compose new mail from existing
This adds a new action to compose a new mail from an existing
one. Before this patch, there was no facility, to finish a draft and
send it. The new action allows to select an old draft, edit it as if it
were a new e-mail with pre-filled fields and body and sent it.

Part of: https://github.com/purebred-mua/purebred/issues/119