Index mails if we keep the draft
Notmuch added new functionality to index individual mails. This patch
uses this functionality in order to index drafts. When keeping a draft,
we store it in our Maildir and index it with notmuch. That way we can
later search for it (e.g. tag:draft), find it and continue editing and
sending it.

The patch updates the version pins as well as the notmuch nix
derivation.