Save and tag sent mail
When sending mail, we want the sent mail indexed and stashed away in our
maildir. Before this patch the mail got sent and got removed from
disk. However who does not want to go through their old sent mails and
dwell in their past glory of having written great pieces of e-mails?

Ok, that is contrived but we need a way to keep sent mail as a reference
in case we need to go back to know what we formerly sent to people. This
patch writes the sent mail into the 'Sent' folder of the local maildir
and indexes it as 'sent'. The latter tag can be customized.