Select last mail if all mail is read
When opening old threads, most of the email in the thread is read.
Most threads on mailing lists are discussions. The last mail in the
thread typically corresponds to the last state of the
discussion. Sometimes a consensus or sometimes just where we left of in
the discussion.

Purebred picked the first unread mail, which is important if there are
unread mails. However Purebred also picked the first read mail, if all
mails were read. This resulted that the user had to close the mail,
scroll down and open the last mail.

This patch selects the last mail by default if the entire thread is
read to avoid the additional key strokes to be made.

Fixes https://github.com/purebred-mua/purebred/issues/205