Render error string centered and fill background
We want the error to be displayed across the full statusbar and wrap
into another row if it is too long. If the error message is too short,
just fill the line with the error colour to give the impression that the
status bar is now showing an error.

The problem with the current code is, that `fillLine` actually doesn't
work as anticipated. The strWrap will create a widget which is fixed in
terms of it's vertical expansion. If the error message is short, it will
only render an error background for the text and the background will
shine through for the rest of the status bar.

If we use `fillLine` it will actually introduce an additional greedy
widget which will expand for the rest of the columns. But since it is
greedy, it will cause the error widget to cause the error message to
wrap due to it's calculation of remaining space.

We really want something which works with both scenarios: fill the rest
of space if the error message is short. Don't introduce arbitrary wraps
if the error message is long.

Using hCenter will exactly produce this, since it calculates the
remaining space left and right and just fill that. If there is no space
left it won't introduce any arbitrary wraps due to the greedyness of
another widget.