Allow stacked rendering of widgets
Brick supports rendering of stacked widgets. For example, Brick can
render a dialog in front of the main UI.

Before this patch, Purebred rendered widgets in a single list onto the
screen. The widgets were rendered vertically from top to
bottom. However, if we wanted to render a dialog in front of the UI it
was rendered at the top and the rest of the widgets used the remaining
height showing them a little bit squished.

This patch changes this by adding an additional abstraction. Instead of
a single list of widgets, purebred allows lists of widgets. Layers can
not be set to hidden or visible, but the containing widgets can.

Each layer renders it's visible widgets resulting in a stacked, rendered
terminal image.

Fixes https://github.com/purebred-mua/purebred/issues/297