Set maildir to actually *the* maildir
In our setup, we set the envMaildir binding to the temp directory in
which we copy the maildir. Since the reset of the setup operated in
there, tests passed. However it would be confusing, because the binding
suggests to be pointing to the Maildir, not the parent directory.

In order to avoid confusion, actually bind the Maildir.