Ability to search for text in mails
Finding substrings in the displayed mail body is something users do. We
could pipe it to less and search there, but that is a very cumbersome
approach.

In order to search for substrings in the mail body, this patch
introduces an additional indirection between the parsed mail and what we
display. This patch adds the first take on a currently very basic AST of
the mail body.

This patch adds the ability to search for sub strings in the abstract
mailbody. This comes at a cost however which is specific to the terminal
library we're using.

Brick provides access to the terminal width only during
rendering. During parsing and searching for sub strings we have no
information available how much text and therefore which line the match
is located at. In order to circumvent this, this patch removes the
ability to consume the available width of the terminal for all body
text. Using a fixed width, we have now the ability to number all lines
at the time we parse the mail body into an AST and therefore are able to
know at which position - horizontally and vertically - the match is.

This is only temporary, since it is not important to know where exactly
the match in the text is. The importance of the line number is only
needed in order to calculate the absolute scrolling height.

Catering for letting text reflow to the full width will be deferred on
to a later card.

Fixes https://github.com/purebred-mua/purebred/issues/279