Add documentation
This adds documentation for most of our functions and the purpose of
those in order to improve general orientation for developers throughout
Purebred.

I made small adjustments to function placements as well as removed dead
code.

The API documentation is by no means complete.

This is related to #137.

Fixes https://github.com/purebred-mua/purebred/issues/324