Piped output (Autoview)
This adds functionality to pipe output through an external program and
capture it's output for display. This allows searching and/or viewing
mails with non-text entities like... HTML.

While it has "Autoview" in the name, this patch isn't delivering what
mutt or any other mail provides. Namely showing the content of all
attachments if they can be piped through an external program. For this,
there are several outstanding things to be implemented like:

a) Extend the AST to actually represent attachments
b) Try to traverse these attachments in order to produce text output
c) Extend the search functionality, most specifically the `ScrollSteps`
   to allow searching through the AST body including the output of the
   attachments.

The entire reason why this is put up for a pull request is the fact that
it already provides some relieve to show non-text mails, especially in
an environment using heavily HTML mails for example.