Reset state of the confirm dialog
When aborting a composition of an e-mail, the confirmation dialog
remembered it's last choice. This can be bad, because it can lead to
accidental deletion of drafts depending on a previous choice. Always
default to 'Keep' no matter what the last choice was.

Fixes https://github.com/purebred-mua/purebred/issues/332