Always position cursor at EOL
The cursor is positioned at the beginning of the line when editing the
fields for an e-mail we're replying to. That is cumbersome to edit,
since most times we want to append, not insert.

This patch inserts the corresponding text into the fields and by that
positions the cursor to the end of the line, instead of explicitly
moving it by calling gotoEOL.

Fixes  https://github.com/purebred-mua/purebred/issues/333