Fix intermittend failing test
This test was failing now and then and the hypothesis causing the
failure is the thread traversing the spine of the list of threads. This
thread most likely evaluates the length of the displayed threads very
quickly. In fact there is only one thread to display, so the length is
most likely returned so quick, that the UI has not been repainted.

Last screen captures from tmux still show the last entry selected, even
though the entry itself is gone.

This patch asserts on the inverse that, the entry (the screen shot
attachment) is gone from the screen. This should be enough to basically
assert that the UI has been repainted.

Fixes https://github.com/purebred-mua/purebred/issues/327