Fix regression of packages
The b698d93f22e7f7f469b7ed25f2c1df9319df4868 commit regressed our
packages in that: it still built purebred, but it had to be split in
such a manner that when a user installs 'purebred', the development
packages come with it. This is important, since the configuration needs
to be recompiled with 'Purebred'.

This patch fixes this problem, but re-introducing the subpackage
structure again.