Change equality test for upsert
The equality test to determine whether a new MIMEMessage (body) is equal
to the same MIMEMessage we have selected used to compare the filename
header. However this test to basically determine whether we updated a
single message body isn't sufficient when it is possible to encapsulate
MIME messages. They will have the same filename header: which is
empty (both are inline attachments). The result of this is, that once we
add an encapsulated message, the Purebred drops us into the encapsulated
message as if we would be editing a composed mail.

This patch only checks whether the headers are equal. When creating new
message bodies, the headers should always be equal while the body
changes. Same thing comparing encapsulated vs new mail bodies: the
headers should be different (just not the particular filename header we
had before).