add --debug=FILE option
Add an option to write debug output to a file.  No logging is
actually done (yet).  No timestamps, levels or other formatting
(yet).  This was just hacked up in a jiffy and if we want something
more sophisticated, we should evaluate established solutions e.g.
co-log.

The log is stored as a sink action (Text -> IO ()) in the AppState.
If no log file is specified via the --debug=FILE option, this is a
no-op.  Otherwise we create a TQueue, the sink puts messages in the
queue, and we spawn a thread to read from the queue and write the
messages to a file.  We don't have any special handling of IO
exceptions and in fact, I don't even know what happens if e.g. the
file isn't writable, it's a normal file, disappears during use, etc.
I guess I'll cross those bridges later.