refactor Action to use StateT
Using 'StateT AppState (EventM n)' for 'Action' definitions improves
the readability (and writeability) of actions.  It also increases
the flexibility (or abstractness) of what data types actions return.

To get all this to work we had to add some orphan 'MonadThrow',
'MonadCatch' and 'MonadMask' instances for '(EventM n)'.  I hope to
land these upstream in brick.