fix instance Focusable 'Threads 'ComposeFrom
The old implementation (which was factored out to a top-level
function 'focusComposeFrom' contained the expression:

  if nullOf (asCompose . cAttachments) s
      then ...

This was erroneously refacted as:

  l <- use (asCompose . cAttachments)
  if null l
    then ...

The latter is testing whether the attachment list is empty, whereas
the former is testing whether the optic (asCompose . cAttachments)
has zero targets.  In the case of an Iso, Lens or Getter, there is
always exactly one target, hence 'nullOf' always returns 'False'!
The true branch of the old implementation is unreachable, but I
introduced a bug (that made it reachable) during the refactor.

To fix, remove the conditional and replace the whole conditional
with the false branch.